It's great to see the progress made on this project. The journey from dealing with unfamiliar technologies likegit PHP OOP and Chart.js to implementing a 3-tier architecture and learning debugging techniques is commendable. The effort put into creating a mockup with AdminLTE to visualize the final project is a good practice.
I appreciate the dedication to creating the first part of the project, which includes a dedicated section for the teacher to manage skills. This is an important step, and it's great to see it in action.

If you have specific questions or need assistance with any aspect of the project, feel free to ask. Additionally, if there are any challenges or specific areas you'd like feedback on, please don't hesitate to share.

Looking forward to the continued progress on this project!

Best regards,
imrane sarsri